refactor(slack-ui): extract Slack channel detail into provider components
